+++ changes.md
@@ -1,2 +1,4 @@
 
 - check latest toolchain version before moon upgrade and add a `--force` flag to force upgrade
+
+- add `--build-graph` flag to generate build graph for build | check | test | bundle